% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{ClustersModel-class}
\alias{ClustersModel-class}
\title{Reference class to represent a clustering model}
\description{
Reference class to represent a clustering model
}
\section{Fields}{

\describe{
\item{\code{clusters}}{list with the differente clusters with their parameters}

\item{\code{fit}}{assigned clusters to the data}
}}


