% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testEvolutiveDynamicConfluenceFastKmeans.R
\name{testEvolutiveDynamicConfluenceFastKmeans}
\alias{testEvolutiveDynamicConfluenceFastKmeans}
\title{Evolving Clustering Dynamic Weights
K-means algorithm}
\usage{
testEvolutiveDynamicConfluenceFastKmeans(file, samplesToInitMeta,
  parameters)
}
\arguments{
\item{file}{path to CSV file with the data}

\item{samplesToInitMeta}{number of samples to init the data}

\item{parameters}{parameters to use}
}
\value{
A list with the error, the elapsed time and the final clusterModel
}
\description{
Evolving Clustering
}
\examples{
tf <- tempfile()
iris2 <- iris
iris2$Species <- as.numeric(iris2$Species)
write.table(iris2,tf,row.names=FALSE, col.names=FALSE,sep=",")
samplesToInitMeta=50
parameters<-list(memory=1,KinitMclust=3)
results=testEvolutiveDynamicConfluenceFastEM(tf,samplesToInitMeta,parameters)
results$clusterModel$fit

}
