% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testEvolutiveDynamicConfluenceFastEM.R
\name{testEvolutiveDynamicConfluenceFastEM}
\alias{testEvolutiveDynamicConfluenceFastEM}
\title{Evolving Clustering Dynamic Weights
EM algorithm}
\usage{
testEvolutiveDynamicConfluenceFastEM(file, samplesToInitMeta, parameters)
}
\arguments{
\item{file}{path to CSV file with the data}

\item{samplesToInitMeta}{number of samples to init the data}

\item{parameters}{parameters to use}
}
\value{
A list with the error, the elapsed time and the final clusterModel
}
\description{
Evolving Clustering
}
\examples{
tf = tempfile()
iris2 = iris
iris2$Species = as.numeric(iris2$Species)
iris2 <- iris2[sample(nrow(iris2)),]
write.table(iris2,tf,row.names=FALSE, col.names=FALSE,sep=",")
samplesToInitMeta=50
parameters=list(memory=2,KinitMclust=3)
resultsA=testEvolutiveDynamicConfluenceFastEM(tf,samplesToInitMeta,parameters)
resultsA$clusterModel$fit
plot(iris2[,-5],col=resultsA$clusterModel$fit)
parameters<-list(memory=0.01,KinitMclust=3)
resultsB=testEvolutiveDynamicConfluenceFastEM(tf,samplesToInitMeta,parameters)
resultsB$clusterModel$fit
plot(iris2[,-5],col=resultsB$clusterModel$fit)

}
